Hopefully fix CAN Logging sync issue.
I believe that the issue with the telemetry is the CAN Metadata and CAN
Payload getting out of sync. These are now pushed through the same RT
FIFO which will resolve the issue if I did correctly identify the cause.
Either way, while the cluster is inelegant, this solution should be
slightly quicker.