Resolve rioembeddedcanlvapi.dll dependency issue.
The fix required that the State Machine (Shutdown & MC) state
indications be disconnected from their respective typedefs in the FPGA
Typedef.ctl. These will need to be refreshed if states are added or
removed from these state machines in order to remain accurate. This will
be detailed in the wiki.